% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.ar1.R
\name{cov.ar1}
\alias{cov.ar1}
\title{estimates covariace matrix for AR1 parameters}
\usage{
cov.ar1(ck, nk, rho, sigma2)
}
\arguments{
\item{ck}{number of cluster with each unique cluster size}

\item{nk}{unique cluster sizes}

\item{rho}{estimated rho}

\item{sigma2}{estimated sigma2}
}
\value{
a list with computed materials
}
\description{
estimates covariace matrix for AR1 parameters
}
\author{
Vahid Nassiri
}
