% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.ar1.R
\name{est.ar1}
\alias{est.ar1}
\title{estimates parameters of a model with AR1 covariance structure
for balanced data}
\usage{
est.ar1(C, n, Y, Plot = 1)
}
\arguments{
\item{C}{number of clusters}

\item{n}{cluster sizes}

\item{Y}{the response}

\item{Plot}{if 1 (Default) will plot the third degree polynomial which is 
solved to find the estimates.}
}
\value{
a list with estimated materials
}
\description{
estimates parameters of a model with AR1 covariance structure
for balanced data
}
\author{
Vahid Nassiri
}
