% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.optimal.ar1.R
\name{iterate.optimal.ar1}
\alias{iterate.optimal.ar1}
\title{computing iterated optimal weights}
\usage{
iterate.optimal.ar1(ck, nk, mu.split, var.comp.split, tol)
}
\arguments{
\item{ck}{number of cluster with each unique cluster size}

\item{nk}{unique cluster sizes}

\item{mu.split}{estimated \eqn{\mu} from different splits}

\item{var.comp.split}{estimated variance components (rho and sigma^2)
from different splits}

\item{tol}{the tolerance}
}
\value{
a list with computed materials
}
\description{
computing iterated optimal weights
}
\author{
Vahid Nassiri
}
